further improvement environment_helper file
